ニックネーム周り修正
